Fix the tag deploy step
The `Makefile` is now autonomous to determine the git tag and the ref,
making this process independant from the CI tool.

It utilizes the command line `git` to parse the current "ref".

The algorithm is the following:

- If the current git ref points to a tag, then the variables `GIT_TAG`
and `GIT_REF` are set to this tag's value.
- Otherwise the variable `GIT_TAG` is undefined, and the variable
`GIT_REF` is set to the current's branch

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>