Add optimise support
Without this, the -a optimise on asciidoctor-pdf will fail with
the following message:

asciidoctor: WARNING: optional gem 'rghost' is not available.
Functionality disabled.

This is a problem, as it prevents the one-liners optimisations,
and requires the users to use a separate gs call.

This fixes it by ensuring rghost is present in the container and
adds tests to guarantee it is working in the future.

Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard-prestataire@ca-gip.fr>